Shared read routines moved to separate file
