+++ src/remove.cpp
@@ -11,6 +11,7 @@
 #include <unordered_map>
 
 #include "fileroutines.h"
+#include "readroutines.h"
 #include "search.h"
 #include "stats.h"
 #include "seq.h"
@@ -41,75 +42,6 @@ void build_patterns(std::ifstream & kmers_f, int polyG, std::vector <std::pair <
     if (polyG) {
         patterns.push_back(std::make_pair(std::string(polyG, 'G'), Node::Type::polyG));
         patterns.push_back(std::make_pair(std::string(polyG, 'C'), Node::Type::polyC));
-    }
-}
-
-/*! \brief Given a read sequence, calculate its DUST score
- *
- *  \param[in]  read    a read sequence
- *  \param[in]  k       the DUST algorithm parameter
- *  \return             the DUST score
- *
- *  \remark For more information on the DUST score, please check the following
- *  paper:
- *  Morgulis, Aleksandr, E. Michael Gertz, Alejandro A. Schäffer, and Richa
- *  Agarwala. "A fast and symmetric DUST implementation to mask low-complexity
- *  DNA sequences." *Journal of Computational Biology* 13, no. 5 (2006): 1028-1040.
- */
-double get_dust_score(std::string const & read, int k)
-{
-    std::unordered_map <int, int> counts;
-    static std::unordered_map <char, int> hashes = {{'N', 1},
-                                          {'A', 2},
-                                          {'C', 3},
-                                          {'G', 4},
-                                          {'T', 5}};
-    unsigned int hash = 0;
-    unsigned int max_pow = pow(10, k - 1);
-    for (auto it = read.begin(); it != read.end(); ++it) {
-        char c = (*it > 96) ? (*it - 32) : *it;
-        hash = hash * 10 + hashes[c];
-        if (it - read.begin() >= k - 1) {
-            ++counts[hash];
-            hash = hash - (hash / max_pow) * max_pow;
-        }
-    }
-    double score = 0;
-    double total = 0;
-    for (auto it = counts.begin(); it != counts.end(); ++it) {
-        score += it->second * (it->second - 1) / 2;
-        total += score;
-    }
-//    std::cout << (total / (read.size() - k + 1)) << std::endl;
-    return (total / (read.size() - k + 1));
-}
-
-/*! \brief Check a read against patterns
- *
- *  \param[in]  read        a read sequence
- *  \param[in]  root        a root of the trie structure used for string matching
- *  \param[in]  patterns    a vector of patterns
- *  \param[in]  length      the read length threshold
- *  \param[in]  dust_k      the DUST algorithm parameter
- *  \param[in]  dust_cutoff the DUST score threshold
- *  \param[in]  errors      the number of resolved mismatches between a read and 
- *                          a pattern
- *  \return                 the read type
- */
-ReadType check_read(std::string const & read, Node * root, std::vector <std::pair<std::string, Node::Type> > const & patterns,
-                    unsigned int length, int dust_k, int dust_cutoff, int errors)
-{
-    if (length && read.size() < length) {
-        return ReadType::length;
-    } 
-    if (dust_cutoff && get_dust_score(read, dust_k) > dust_cutoff) {
-        return ReadType::dust;
-    }
-    
-    if (errors) {
-        return (ReadType)search_inexact(read, root, patterns, errors);
-    } else {
-        return (ReadType)search_any(read, root);
     }
 }
